feat: auto-recovery startup scan, data visibility model & remove Sync MAST Files
- Extract scan-and-import logic from DataManagementController into
  DataScanService (IDataScanService / DataScanService)
- Add StartupScanBackgroundService that runs once on backend startup:
  scans disk for FITS files, imports missing records, enqueues
  thumbnail generation for records missing thumbnails
- Change IsPublic default to true on JwstDataModel and MastImportRequest
  (MAST data is public NASA data, uploads default public)
- MosaicService: inherit IsPublic from source records instead of
  hardcoding false (all-public sources → public, any private → private)
- Remove "Sync MAST Files" button from frontend (DashboardToolbar,
  JwstDataDashboard) — no longer needed with automatic startup scan
- Update docs (key-files, quick-reference, architecture)
- Fix test assertion for new IsPublic default
